fix(exec): report unavailable PTYs on Windows
